Adding callMockEvent function, updating tests
